Refactor resource data type store
