+++ stores/index.ts
@@ -1,7 +1,7 @@
 export { default as useProjectKeyStore } from './useProjectKeyStore';
 export { default as useResourceDataStore } from './useResourceDataStore';
-export { default as useResourceDataTypeStore } from './useResourceDataTypeStore';
 
 export { default as useApi } from './useApi';
 export { default as useProject } from './useProject';
+export { default as useResourceDataType } from './useResourceDataType';
 export { default as useResourceModel } from './useResourceModel';
